SUMMARY:  Another passle of changes to the Trilinos 10.0 branch to make that angry Kallikrates MS test machine happy.
     BUGZILLA BUG #(s):

     CHECKLIST
     ---------

     1) Did you 'cvs add' all new files and directories [Y:N:N/A]?  n/a

     2) Did you add new source files to a CMakeLists.txt file [Y:N:N/A]?  n/a

     3)  Did you modify or add any source or build system files?  no


     DESCRIPTION OF CHANGES:

     Summary says it all.  All serial/mpi tests pass on my machine.

(checklist version 2009-07-14)

+++ packages/ml/src/Utils/ml_memory.c
@@ -1021,7 +1021,7 @@ int ML_MaxAllocatableSize()
 {
   int* junk = NULL;
   long long int upper, mid, lower;
-  long long int ml_total_mem;
+  size_t ml_total_mem;
 
   ml_total_mem = 10000;
   while ( (junk = (int*) malloc(sizeof(int)*ml_total_mem)) != NULL) {
